DPO_ZymCTRL/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
aniso8601==10.0.1
annotated-types==0.7.0
ansi2html==1.9.2
antlr4-python3-runtime==4.9.3
anyio==4.9.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
attrs==23.2.0
Authlib==1.5.2
babel==2.17.0
beautifulsoup4==4.13.4
biopython==1.85
biotite==0.41.2
biotraj==1.2.2
bleach==6.2.0
blinker==1.9.0
bokeh==3.7.2
boto3==1.38.2
botocore==1.38.2
Brotli==1.1.0
captum==0.8.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.21.0
cloudpickle==3.1.1
cmake==3.26.4
comm==0.2.2
contourpy==1.3.2
cryptography==44.0.2
cycler==0.12.1
datasets==3.5.1
debugpy==1.8.14
decorator==5.2.1
defusedxml==0.7.1
dill==0.3.8
docker==7.1.0
docker-pycreds==0.4.0
dparse==0.6.4
einops==0.8.1
entrypoints==0.4
enum-compat==0.0.3
executing==2.2.0
fair-esm @ git+https://github.com/facebookresearch/esm.git@2b369911bb5b4b0dda914521b9475cad1656b2ac
fastapi==0.115.12
fastjsonschema==2.21.1
filelock==3.16.1
Flask==3.1.0
Flask-RESTful==0.3.10
fonttools==4.57.0
fqdn==1.5.1
frozenlist==1.6.0
fsspec==2025.3.0
gitdb==4.0.12
GitPython==3.1.44
google-pasta==0.2.0
gym==0.26.2
gym-notices==0.0.8
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.30.2
idna==3.10
imageio==2.37.0
importlib-metadata==6.11.0
iniconfig==2.1.0
ipykernel==6.29.5
ipython-genutils==0.2.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.6
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.4.2
json5==0.12.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==7.4.9
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.1
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.14
kiwisolver==1.4.8
lightning-utilities==0.14.3
llvmlite==0.44.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==4.0.0
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.3
mock==4.0.3
mpmath==1.3.0
msgpack==1.1.0
msgpack-numpy==0.4.8
multidict==6.4.3
multiprocess==0.70.16
narwhals==1.36.0
nbclassic==1.3.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
notebook==6.5.7
notebook_shim==0.2.4
numba==0.61.2
numpy==1.26.4
nvgpu==0.10.0
omegaconf==2.3.0
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.7
pluggy==1.5.0
pox==0.3.6
ppft==1.7.7
prettytable==3.16.0
prometheus_client==0.21.1
prompt_toolkit==3.0.51
propcache==0.3.1
protobuf==5.29.4
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==19.0.1
pybind11==2.13.6
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pyfunctional==1.5.0
pygame==2.6.1
Pygments==2.19.1
pynvml==12.0.0
pyparsing==3.2.3
pytest==8.3.5
python-dateutil==2.9.0.post0
python-json-logger==3.3.0
pytorch-lightning==2.5.1.post0
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.4.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.24.0
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
s3fs==0.4.2
s3transfer==0.12.0
safetensors==0.5.3
safety-schemas==0.0.14
sagemaker==2.243.3
sagemaker-core==1.0.29
schema==0.7.7
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
Send2Trash==1.8.3
sentencepiece==0.2.0
sentry-sdk==2.27.0
setproctitle==1.3.5
setuptools==79.0.1
shap==0.47.2
shellingham==1.5.4
six==1.17.0
slicer==0.0.8
smclarify==0.5
smdebug-rulesconfig==1.0.1
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
stack-data==0.6.3
starlette==0.46.2
sympy==1.13.1
tabulate==0.9.0
tblib==3.1.0
tenacity==9.1.2
termcolor==3.0.1
terminado==0.18.1
threadpoolctl==3.6.0
tinycss2==1.4.0
tokenizers==0.21.1
tomlkit==0.13.2
torch==2.5.1
torch-geometric==2.6.1
torch-model-archiver==0.12.0
torch-workflow-archiver==0.2.15
torchaudio==2.5.1
torchmetrics==1.7.1
torchserve==0.11.1
torchtext==0.6.0
torchvision==0.20.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.48.1
triton==3.1.0
typer==0.15.2
types-python-dateutil==2.9.0.20241206
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
uri-template==1.3.0
urllib3==2.4.0
uvicorn==0.34.2
wandb==0.19.10
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
wheel==0.45.1
widgetsnbextension==4.0.14
xxhash==3.5.0
xyzservices==2025.1.0
yarl==1.20.0
zipp==3.21.0
zstd==1.5.6.7
python-Levenshtein